Revert "TS migration and login changes"
