+++ components/modal.module.css
@@ -31,7 +31,4 @@
     width: 100%;
     height: 100%;
     background-color: rgba(0, 0, 0, 0.6);
-    /* background-image: url("../public/images/noise.png");
-    background-size: cover;
-    opacity: 0.5; */
 }
