+++ lib/mongodb.js
@@ -32,8 +32,7 @@ async function dbConnect() {
             bufferCommands: false,
         }
 
-        const uri = MONGODB_URI || '';
-        cached.promise = mongoose.connect(uri, opts).then((mongoose) => {
+        cached.promise = mongoose.connect(MONGODB_URI, opts).then((mongoose) => {
             return mongoose;
         });
     }
